// https-apps-api/packages/api/src/properties/properties.schemas.ts
import * as z from "zod";

export const propertiesPostRequestSchema = z.object({
    targetId: z.number(),
    targetType: z.enum(['node','asset','workflowSeries','appSeries','dbCluster','job']),
    properties: z.any() // fixme: set to any object


})
export type PropertiesPostReqBody = z.infer<typeof propertiesPostRequestSchema>